+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_task.py
@@ -50,7 +50,7 @@ from ..errors import (
     DynamicSidecarError,
     DynamicSidecarNotFoundError,
 )
-from ._core._core import observing_single_service
+from ._core._observer import observing_single_service
 
 logger = logging.getLogger(__name__)
